# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Bank of Blockchain API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user info
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /balances/me:
    get:
      operationId: getMyBalances
      tags: [balances]
      summary: Get current user balances
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Balances
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalancesResponse"
  /balances/{userId}:
    get:
      operationId: getUserBalances
      tags: [balances]
      summary: Get user balances (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Balances
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalancesResponse"
    patch:
      operationId: updateUserBalances
      tags: [balances]
      summary: Update user balances (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBalancesRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalancesResponse"
...
  /bank-transfers:
    get:
      operationId: getBankTransfers
      tags: [bank-transfers]
      summary: Get bank wire transfers
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Bank transfers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankTransfersListResponse"
    post:
      operationId: createBankTransfer
      tags: [bank-transfers]
      summary: Request a bank wire transfer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBankTransferRequest"
      responses:
        "201":
          description: Bank transfer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankTransfer"
  /bank-transfers/{id}:
    get:
      operationId: getBankTransferById
      tags: [bank-transfers]
      summary: Get bank transfer by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bank transfer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankTransfer"
  /bank-transfers/{id}/validate:
    post:
      operationId: validateBankTransfer
      tags: [bank-transfers]
      summary: Validate bank transfer (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Validated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /bank-transfers/{id}/reject:
    post:
      operationId: rejectBankTransfer
      tags: [bank-transfers]
      summary: Reject bank transfer (admin)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectRequest"
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /bank-accounts:
    get:
      operationId: getBankAccounts
      tags: [bank-accounts]
      summary: Get user bank accounts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Bank accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BankAccount"
    post:
      operationId: createBankAccount
      tags: [bank-accounts]
      summary: Add a bank account
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBankAccountRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankAccount"
  /bank-accounts/{id}:
    delete:
      operationId: deleteBankAccount
      tags: [bank-accounts]
      summary: Delete a bank account
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        country:
          type: string
        dateOfBirth:
          type: string
      required:
        - email
        - password
        - firstName
        - lastName
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
      required:
        - token
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        country:
          type: string
        dateOfBirth:
          type: string
        role:
          type: string
          enum: [user, admin]
        status:
          type: string
          enum: [pending, active, suspended]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - firstName
        - lastName
        - role
        - status
        - createdAt
...
    Balances:
      type: object
      properties:
        eur:
          type: number
        usd:
          type: number
        btc:
          type: number
      required:
        - eur
        - usd
        - btc
    BalancesResponse:
      type: object
      properties:
        userId:
          type: integer
        balances:
          $ref: "#/components/schemas/Balances"
      required:
        - userId
        - balances
    UpdateBalancesRequest:
      type: object
      properties:
        eur:
          type: number
        usd:
          type: number
        btc:
          type: number
...
    RejectRequest:
      type: object
      properties:
        reason:
          type: string
    BankTransfer:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
        currency:
          type: string
          enum: [EUR, USD]
        beneficiaryName:
          type: string
        iban:
          type: string
        bic:
          type: string
        bankName:
          type: string
        reference:
          type: string
        status:
          type: string
          enum: [pending, processing, completed, rejected]
        rejectionReason:
          type: string
        transactionId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        user:
          type: object
          properties:
            firstName:
              type: string
            lastName:
              type: string
            email:
              type: string
      required:
        - id
        - userId
        - amount
        - currency
        - beneficiaryName
        - iban
        - status
        - createdAt
    BankTransfersListResponse:
      type: object
      properties:
        transfers:
          type: array
          items:
            $ref: "#/components/schemas/BankTransfer"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - transfers
        - total
        - page
        - limit
    CreateBankTransferRequest:
      type: object
      properties:
        amount:
          type: number
        currency:
          type: string
          enum: [EUR, USD]
        beneficiaryName:
          type: string
        iban:
          type: string
        bic:
          type: string
        bankName:
          type: string
        reference:
          type: string
        bankAccountId:
          type: integer
      required:
        - amount
        - currency
        - beneficiaryName
        - iban
...
    BankAccount:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        beneficiaryName:
          type: string
        iban:
          type: string
        bic:
          type: string
        bankName:
          type: string
        country:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - beneficiaryName
        - iban
        - createdAt
    CreateBankAccountRequest:
      type: object
      properties:
        beneficiaryName:
          type: string
        iban:
          type: string
        bic:
          type: string
        bankName:
          type: string
        country:
          type: string
      required:
        - beneficiaryName
        - iban
...