# lib/api-zod/src/generated/types/bankTransfer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Bank of Blockchain API
 * OpenAPI spec version: 0.1.0
 */
import type { BankTransferCurrency } from "./bankTransferCurrency";
import type { BankTransferStatus } from "./bankTransferStatus";
import type { BankTransferUser } from "./bankTransferUser";

export interface BankTransfer {
  id: number;
  userId: number;
  amount: number;
  currency: BankTransferCurrency;
  beneficiaryName: string;
  iban: string;
  bic?: string;
  bankName?: string;
  reference?: string;
  status: BankTransferStatus;
  rejectionReason?: string;
  transactionId?: number;
  createdAt: Date;
  updatedAt?: Date;
  user?: BankTransferUser;
}